rename api.ts to deprecated.ts
