+++ src/api/deprecated.ts
@@ -1,3 +1,7 @@
+/**
+ * This files the obsolete API. Use the new API in provider.ts instead.
+ */
+
 import { BlockSubpathResult, HeadingSubpathResult, PluginManifest, TFile } from 'obsidian';
 import MathLinks from '../main';
 import { informChange } from 'src/utils';
